MDMovementReason - index Int16 -> short

+++ gip.mes.datamodel/Enums/MovementReasonsEnum.cs
@@ -32,19 +32,19 @@ namespace gip.mes.datamodel
     {
         public ACValueListMovementReasonsEnum() : base("MovementReasons")
         {
-            AddEntry((int)MovementReasonsEnum.Adjustment, "en{'Adjustment'}de{'Korrektur'}");
-            AddEntry((int)MovementReasonsEnum.ZeroStock, "en{'Zero Stock'}de{'Nullbestand'}");
-            AddEntry((int)MovementReasonsEnum.Enabling, "en{'Enabling'}de{'Freigabe'}");
-            AddEntry((int)MovementReasonsEnum.Blocking, "en{'Blocking'}de{'Sperrung'}");
-            AddEntry((int)MovementReasonsEnum.Consumption, "en{'Consumption'}de{'Verbrauch'}");
-            AddEntry((int)MovementReasonsEnum.Production, "en{'Production'}de{'Herstellung'}");
-            AddEntry((int)MovementReasonsEnum.GoodsReceipt, "en{'Goods Receipt'}de{'Wareneingang'}");
-            AddEntry((int)MovementReasonsEnum.GoodsIssue, "en{'Goods Issue'}de{'Warenausgang'}");
-            AddEntry((int)MovementReasonsEnum.Relocation, "en{'Relocation'}de{'Umlagerung'}");
-            AddEntry((int)MovementReasonsEnum.Inventory, "en{'Inventory'}de{'Inventur'}");
-            AddEntry((int)MovementReasonsEnum.ConsumptionWithoutBalance, "en{'Consumption Without Balance'}de{'Verbrauch ohne Bilanz'}");
-            AddEntry((int)MovementReasonsEnum.ProductionWithoutBalance, "en{'Production Without Balance'}de{'Herstellung ohne Bilanz'}");
-            AddEntry((int)MovementReasonsEnum.Reject, "en{'Reject'}de{'Ausschuss'}");
+            AddEntry((short)MovementReasonsEnum.Adjustment, "en{'Adjustment'}de{'Korrektur'}");
+            AddEntry((short)MovementReasonsEnum.ZeroStock, "en{'Zero Stock'}de{'Nullbestand'}");
+            AddEntry((short)MovementReasonsEnum.Enabling, "en{'Enabling'}de{'Freigabe'}");
+            AddEntry((short)MovementReasonsEnum.Blocking, "en{'Blocking'}de{'Sperrung'}");
+            AddEntry((short)MovementReasonsEnum.Consumption, "en{'Consumption'}de{'Verbrauch'}");
+            AddEntry((short)MovementReasonsEnum.Production, "en{'Production'}de{'Herstellung'}");
+            AddEntry((short)MovementReasonsEnum.GoodsReceipt, "en{'Goods Receipt'}de{'Wareneingang'}");
+            AddEntry((short)MovementReasonsEnum.GoodsIssue, "en{'Goods Issue'}de{'Warenausgang'}");
+            AddEntry((short)MovementReasonsEnum.Relocation, "en{'Relocation'}de{'Umlagerung'}");
+            AddEntry((short)MovementReasonsEnum.Inventory, "en{'Inventory'}de{'Inventur'}");
+            AddEntry((short)MovementReasonsEnum.ConsumptionWithoutBalance, "en{'Consumption Without Balance'}de{'Verbrauch ohne Bilanz'}");
+            AddEntry((short)MovementReasonsEnum.ProductionWithoutBalance, "en{'Production Without Balance'}de{'Herstellung ohne Bilanz'}");
+            AddEntry((short)MovementReasonsEnum.Reject, "en{'Reject'}de{'Ausschuss'}");
         }
     }
 #endif
